refactoring: 디렉토리 구조 수정
도메인별로 묶어줌
