+++ src/main/java/com/example/e2ekernelengine/blog/util/BlogOwnerType.java
@@ -1,4 +1,4 @@
-package com.example.e2ekernelengine.entity;
+package com.example.e2ekernelengine.blog.util;
 
 public enum BlogOwnerType {
 	INDIVIDUAL, COMPANY
