+++ src/main/java/com/example/e2ekernelengine/feed/dto/request/FeedSearchRequest.java
@@ -1,0 +1,6 @@
+package com.example.e2ekernelengine.feed.dto.request;
+
+public class FeedSearchRequest {
+	private String keyword; // 검색어
+	private Integer maxResults; // 최대 결과 수 제한
+}
