+++ src/main/java/com/example/e2ekernelengine/feed/dto/response/FeedSearchResponse.java
@@ -13,4 +13,9 @@ public class FeedSearchResponse {
 	private final String feedTitle;
 	private final String feedContent;
 
+	// 정적 팩토리 메서드를 사용하여 객체 생성
+	public static FeedSearchResponse create(String feedId, String feedUrl, String feedTitle, String feedContent) {
+		return new FeedSearchResponse(feedId, feedUrl, feedTitle, feedContent);
+	}
+
 }
