+++ src/main/java/com/example/e2ekernelengine/feed/service/FeedService.java
@@ -1,0 +1,6 @@
+package com.example.e2ekernelengine.feed.service;
+
+public interface FeedService {
+
+}
+
