+++ src/main/java/com/example/e2ekernelengine/user/util/UserStatusType.java
@@ -1,4 +1,4 @@
-package com.example.e2ekernelengine.entity;
+package com.example.e2ekernelengine.user.util;
 
 public enum UserStatusType {
 	ACTIVE("active"),
